Workaround for "Multiple conflicting contents for sourcemap..."
See https://github.com/vuejs/rollup-plugin-vue/issues/238

In this case, Rollup throws a hard error and aborts. This is highly disruptive for all that the issue when skipping would be a missing or corrupt sourcemap! Instead, let's just skip this particular iteration. It's not the prettiest workaround, but it does have the benefit of actually making us productive again. Woo!